cryptohome: Add TimeToPerformEphemeralMount metric.
cryptohome is currently reporting the overall time taken in
MountGuestCryptohome, as well as the time taken specifically to fork,
exect the mount helper, perform an ephemeral mount, and signal the main
cryptohome process.

In order to quantify the overhead of the out-of-process mounter, report
the amount of time taken *inside* the helper process just to perform
the ephemeral mount. This, compared with the
TimeToPerformOOPMountOperation metric should give a good indication of
the OOP overhead.

BUG=chromium:1074735
TEST=Build, log in, check chrome:histograms.

Change-Id: Ife21633e8abfaa197cce8d487f2e83cc85b4f4b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2261052
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>