lorgnette: fix bug in color mode parsing
Lorgnette is using ColorMode_Parse to parse color mode values, but this
is failing since the values are not stringified proto enums but rather
the constants specified in dbus-constants.h.

Add a helper method that parsees the dbus strings appropriately and fix
ScanImage.

Update unit tests so that they would catch a similar bug in the future.

BUG=None
TEST=unit tests

Change-Id: I72872b5e0546b5a5e9e264f563d1f1aad76f5d61
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2264841
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>