cryptohome: Add fscrypt policy awareness to functions
In preparation for adding fscrypt v2 policies, add a key reference
which holds the policy version and a key reference. Depending on
the policy, the key reference is either the key descriptor generated
in userspace and passed to the kernel (v1 policies) or the identifier
generated once the key is added to the filesystem.

BUG=b:152332673, chromium:777613
TEST=unittests

Change-Id: Ifbeada730f71adb76e542126d0abf0a8d51691f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2274708
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>