diagnostics: Launch cros_healthd and connect up the executor
Start up cros_healthd and send an executor invitation to it. From the
cros_healthd side, receive the invitation and add the executor to the
context, so it can be used as needed in cros_healthd.

BUG=chromium:1085464
TEST=FEATURES=test emerge-nami diagnostics
Run '/usr/bin/cros_healthd --executor' and see that it starts both the
executor and cros_healthd without any errors in the logs.

Change-Id: I0b45a4e8b11e11c2c2aa97991c2961978bb05015
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2268327
Reviewed-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>