foomatic_shell: First implementation of complete mini-shell
foomatic_shell is a mini-shell that will be used by foomatic-rip
instead of default ChromeOS shell. It allows to control scripts
included in some PPD files. This CL contains an implementation of
complete mini-shell functionality built on top of the parser committed
in CL:2198690 and the verifier in CL:2295623.

BUG=chromium:1073063
TEST=on atlas with platform_PrinterPpds.all_test

Change-Id: I57cf05ac2ba58cc5f3a2ec497ff4b672bdb0b4f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2271389
Commit-Queue: Piotr Pawliczek <pawliczek@chromium.org>
Tested-by: Piotr Pawliczek <pawliczek@chromium.org>
Reviewed-by: Kalvin Lee <kdlee@chromium.org>
Reviewed-by: Luum Habtemariam <luum@chromium.org>