shill: net: fix nlmsg_len for RTNL dump messages
Currently, we treat the payload of a RTNL dump message of any type as a
`struct ifinfomsg`, while the kernel may parse it as a different struct,
and this may incur some problems. For example, for a neighbor dump RTNL
message, kernel expects its payload as a `struct ndmsg`, since these two
struct have different length (ndmsg: 12, ifinfomsg: 16), when kernel
parses the message, it will try to parse the attributes from a wrong
start point, and then log some errors like "netlink: xx bytes leftover
after parsing attributes".

This patch corrects the length for the dump messages of all the types,
and adds a unit test for it.

BUG=b:159936699
TEST=`FEATURES="test" emerge-$BOARD shill`;
TEST=manually test on atlas, no more "xx bytes leftover" error in
  /var/log/messages after restart patchpanel.

Change-Id: Ia4a950652d63eb72b6ae1eaaabd087e8a4184951
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2272818
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>