shill: publish technology specific connection failure metrics
We have metrics to find the successful wifi connections
but there is not reliable metrics to identify wifi
network connection failures. There is a metrics for
generic network connection failures but it does not
differentiate between different interfaces like wifi,
cellular or ethernet. Thus there is not proper way to
track network failures when wifi is the interface.

Solution: append the network service failure with technology
as well. This will track network connection failures
happening for a particular interface. The metrics published is
"Network.shill.<technology>.ServiceErrors" in case of wifi
technology it comes out to be
Network.Shill.Wifi.ServiceErrors which was verified in histograms

BUG=chromium:1085300
TEST= - Unit test passed: FEATURES=test emerge-octopus shill
- Call SendServiceFailure method in the happy condition to
trigger the metrics with value kFailureNone (0).
- Verified the metrics in chrome://histograms

Change-Id: Iad1ae1ecc85e24b5d532392c91a32e07905ee753
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2277056
Tested-by: Abhishek Kumar <kuabhs@chromium.org>
Commit-Queue: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>