shill: Service: Reset static ip parameters on Unload
Static IP parameters were not getting reset when a service was
unloaded. This caused the parameters to remain applied when a
Service was reconnected after Unload (forget).

This also adds ServiceTest.StaticIPConfigs to test the behavior,
and converts some mocks in ServiceTest to NiceMock to reduce
spam when running tests.

BUG=chromium:898881
TEST=P2_TEST_FILTER=ServiceTest.StaticIPConfigs cros_workon_make --board=${BOARD} --test shill

Change-Id: I2ef875b72b262dba45d03866d67ca490e851d24b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2276716
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>