hermes: modem_qrtr: Uniformly check for response errors
Prior to this change, ModemQrtr did not have uniform logic in checking
for errors in QMI responses and logging any relevant messages. This
change unifies this logic and also modified qmi2cpp so that
QmiUimCommand can be used either as a uint16_t or a string.

BUG=chromium:1085825
TEST=All unit tests are passing.

Change-Id: Ia7a140da21f59545ae8582fc419d6ee4d455b322
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2281070
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Pavan Holla <pholla@google.com>