diagnostics: add ProcessFetcher
Add a new ProcessFetcher, which fetches information related to a
single process on the device. This fetcher functions differently
from the rest, and a separate instance needs to be created for each
process whose information is desired.

Also add a new Mojo method to request information about a process.
The reason this data isn't included in the normal TelemetryInfo
struct is because returning ProcessInfo for every process on the
device would be a lot of data, potentially being transferred
frequently. Instead, clients will need to request specific PIDs
in this new method.

BUG=chromium:1102518
TEST=FEATURES=test emerge-nami diagnostics

Change-Id: I64b3db906057dadbb3c89772af87f24869c83f50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2283687
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>