patchpanel: Remove ARC sentinel.
ag/12085176 changes the sentinel used on ARC launch so this can be
removed as it is no longer used.

This patch should land after ag/12085176 but it's not strictly necessary
as (a) the network will be up anyway when the wait times out but (b) it
doesn't actually seem to matter to ARC anymore.

BUG=chromium:1102361
BUG=b:160662073
TEST=unit tests

Change-Id: I02c66f326805c9d575ee663d8290f1f2f6b2e0f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2284329
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Garrick Evans <garrick@chromium.org>