vm_tools: Fix crash in ~VmExportOperation
The default destructor would delete the memory containing a hash of
the data written so far before destroying the object representing the
write operation. This could cause a crash if the destruction occurred
in the middle of the operation, because cancelling it would trigger
callbacks that expected to be able to access the hash. Fix this by
adding an explicit destructor that triggers these callbacks before
deleting anything else.

BUG=chromium:1102800
TEST=none

Change-Id: Ia893af1fbc6a7732f1fd513fb7ec3ab92a720d05
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2284882
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>