cryptohome: Update VKK to treat timestamp file read error as non-fatal
Previously, if VKK is unable to read the timestamp file, it'll fail the
entire Load() operation, and causing unlock/mount to fail, ultimately
requiring the user to recreate the vault.

This change update such failure to be only a warning in the log.

BUG=b:160238036
TEST=Remove timestamp files and login

Change-Id: I77bfb356b2938281492c0a53d7ac7b457a40c4f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2284884
Tested-by: John L Chen <zuan@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Commit-Queue: John L Chen <zuan@chromium.org>