tpm_manager: Fix DA Reset even when Lockout is in effect during startup for TPMv1.2
Dictionary Attack reset for TPMv1.2 requires the ownership status to be
available. However, with TPMv1.2, if the dictionary attack lockout is in
effect, it would block getting the ownership status and cause the reset
to fail. This would cause a unit to be stuck in DA lockout for extensive
amount of time, and is not preferable.

BUG=b:155063317
TEST=Manually lockout a device, then restart tpm_managerd to see that DA reset works

Change-Id: I36d114e782b13ac0c282f1bf08d3bc713bfd304f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2284885
Tested-by: John L Chen <zuan@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Commit-Queue: John L Chen <zuan@chromium.org>