system-proxy: Delay call to patchpanel::ConnectNamespace
The network setup for System-proxy is handled by patchpaneld.
Patchpaneld needs to enter the network namespace of the worker process
created by System-proxy in order to take one half of the veth pair into
the host namespace.

The command "nsenter" fails if it's called to early after the
System-proxy worker is started. This CL adds a small delay before
calling patchpanel::ConnectNamespace to avoid this race condition.

BUG=chromium:1095170
TEST=manually on the DUT restart system-proxy and check from the logs
     that the namespace was correctly setup every time:
       for i in {1..20}; do restart ui && sleep 3; done

Change-Id: I60464eb665400d1cd9f97c63f363a88709a866a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2287435
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>