shill: ethernet: CreateService in Ethernet::Start instead of ctor
Prior to this change, the service_ of an Ethernet Device would be
initialized in the constructor and wouldn't change (save for switching
to and from PPPoE mode). This model is actually incorrect because
changing the order of EthernetServices can also change the Ethernet
Device to which it should be associated. For example, if an Ethernet
Device is created and the interface eventually goes down, the Ethernet
Device still holds a pointer to the same EthernetService. If another
Ethernet Device is created, EthernetProvider::CreateService will
repurpose this same EthernetService for the new Device, but the old
Device will still point to the same Service.

The root cause of this is that there's an asymmetry between
EthernetProvider::RegisterService and
EthernetProvider::DeregisterService, where ResetEthernet is called in
DeregisterService, but SetEthernet is never called in RegisterService
(it's done in CreateService, which is called from Ethernet::Start). This
is fine for most use cases since Ethernet is almost always only Start-ed
right after construction and Stop-ed right before destruction, but some
integration tests use the DisableTechnology Manager D-Bus function on
Ethernet, which Stop-s the Device without destroying it.

In this change, we move CreateService to Ethernet::Start. This breaks
the assumption that the service_ of an Ethernet Device always exists, so
add some CHECKs and 'if (service_)' wrappers to make sure we don't hit
any null derefs. In general, this should not be possible since an
Ethernet Device won't be used unless it has been enabled (and thus
Start-ed), but forcing the link up and down (ifconfig eth0 up/down) can
have some weird side affects, e.g. Ethernet::LinkEvent will still be
called and we will start eap_listener_ even if no service_ exists. Since
we didn't handle this case before, this behavior isn't changed here.

BUG=chromium:1065663
TEST=1) DisableTechnology and then EnableTechnology via D-Bus no longer
        crashes shill.
     2) Ethernet8021X tast tests
     3) EthernetStaticIP tast test
     4) SetUSBEthernetMACAddressSource tast test
     5) Unit tests

Change-Id: I4d480562ea6d74da4124e77db6b460011a783b9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2289513
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>