vm_tools: maitred: Set GOTRACEBACK=crash
Currently when our go binaries crash, either due to an internal panic
or an external signal, the go runtime catches this and prevents a core
dump from being created. This is bad because it prevents us from
learning about the failure and debugging it. Set the environment
variable GOTRACEBACK to "crash" to change this behaviour to dumping
core.

It's unclear at this point whether the minidumps that get uploaded
will be useful without further processing, but we can at least do this.

BUG=chromium:1103093
TEST=Made LXD panic
TEST=Killed LXD with SIGABRT

Change-Id: I6724382fa46fe6438084838ef7da647adf9493f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2289727
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>