diagnostics: initialize base::Value to dict
In the latest version of libbase, base::Value checks if the type is a
dictionary before setting keys and values. To use it correctly, it needs
to be initialized as a dictionary for the diagnostics dpsl test
binaries.

BUG=b:161125460
TEST=build VM image and test diagnostics_dpsl_test_* binaries

Change-Id: I6bfdedc75f354c093e303dc03f65353820276216
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2295626
Tested-by: Trent Begin <tbegin@chromium.org>
Auto-Submit: Trent Begin <tbegin@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Paul Moy <pmoy@chromium.org>
Commit-Queue: Oleh Lamzin <lamzin@google.com>