camera: libcamera_connector: Tune camera info generation logic
Using a similar logic found in VCD, we tune the logic we used to
generate camera info. We look at the AE FPS ranges and use those to
derive a list of attainable FPS for any given dimension. This is needed
for devices that support 60fps recording (e.g., atlas).

BUG=chromium:1096218,b:151047930,b:167994590
TEST=Run cros_camera_connector_test on atlas.

Change-Id: Icbd2fa74e26b26c25acd16b1148b377b7a9f1cff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2297130
Reviewed-by: Shik Chen <shik@chromium.org>
Commit-Queue: Jasmine Chen <lnishan@google.com>
Tested-by: Jasmine Chen <lnishan@google.com>