vm_tools: concierge: handle "not supported" error from permission service
Chrome will be temporarily returning "not supported" when attempting to
register Parallels VMs with permission service if both camera and
microphone feature flags are disabled (not to be mixed up with
permissions themselves). In this case we want to pass empty permission
token to Parallels dispatcher/UI so that they can suppress Camera and
Microphone icons in their UI.

This will be removed later when Chrome feature flags are removed.

BUG=chromium:1071872
TEST=Start VMs, verify camera can be toggled.

Change-Id: Idf72ff5a08935a1eeca9489a3a5d5fea5327bc6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2296848
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>