common-mk: Skip tests that rely on gn if gn is not available
We'd like to be able to run the presubmit checks server-side so that
they are enforced consistently. Currently the presubmit check does not
provide a full chroot: https://crbug.com/1078990.

For now we'll skip the tests that require gn if gn is not available.
This is no worse than the current situation because the tests will still
be run locally when presubmit runs, but they will be skipped
server-side.

BUG=chromium:1101555, chromium:1078990
TEST=./common-mk/gnlint_unittest.py

Change-Id: I0e3ebba864152409de676f1c44abcfda68d6b044
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2300708
Tested-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>