ippusb_manager: handle EINTR and log errnos
Add HANDLE_EINTR wrappers around some of the syscalls in socket_manager.
When appropriate, replace LOG statements with PLOG so that errno will be
printed as well.

The purpose of this is to minimize reporting spurious failures and
closing ippusb_manager when we should have just retried a syscall.

BUG=None
TEST=run printer.PrintIPPUSB

Change-Id: I3884ddcb9695b6aefc9ede429cb3ec8e5e3fb66b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2300711
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>