diagnostics: switch from mock to fake for SystemUtilities
The way CpuFetchersTest is written, an initial valid configuration
is set up, with individual tests expected to overwrite the parts
of that configuration that they need to. With a mock, the initial
configuration needs to set an expectation on that mock, and then
tests would need to reset the expectation before placing their
own. This is not obvious, and a better solution would be to have a
fake where no expectations are necessary.

BUG=chromium:1105970
TEST=FEATURES=test emerge-nami diagnostics

Change-Id: I25a3b19145a4f8268e0df7b2c623d5c85d99d1e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2300718
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>