cryptohome: Take Bus in KeyChallengeServiceFactory
Change the KeyChallengeServiceFactory interface to receive a dbus::Bus
parameter instead of a brillo::DBusConnection.

This is a preparation for porting the CheckKey unittests from the Service
class to the UserDataAuth class, since the latter uses dbus::Bus for
referring to the system D-Bus.

This is a pure refactoring CL.

BUG=chromium:1044558
TEST=cros_workon_make cryptohome --test

Change-Id: Id21bab49e172726df845784e26c1dee11a1ffea4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2302896
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: John L Chen <zuan@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>