chromeos-config: use image-name as key if coreboot not present
If we do not have a coreboot firmware build target, then we need a way
to name the firmware bundle of build targets. The image-name should be
the correct field to properly bundle the firmware build targets.

This is needed to allow us to build a zephyr-ec firmware build on a VM
target that does not have a coreboot target.

BUG=b:161555377,b:157245202
TEST=See that betty config without coreboot name does not error

Change-Id: I1ae6443dca29fd007e7312443abf122a3ec24f55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2304542
Tested-by: Jett Rink <jettrink@chromium.org>
Commit-Queue: Jett Rink <jettrink@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>