shill: remove CHECK() from ConfigurePPPoEMode
With the changes in CL:2289513, we can no longer expect that service_
will be non-null when ConfigurePPPoEMode is called. Particularly, it is
called before the Ethernet Device is Start()-ed in Ethernet::Load().
Remove the CHECK at the beginning of the function and allow PPPoE
services to be created before Ethernet::Start() is called. We still
maintain that standard EthernetServices are created only when an
Ethernet Device is Start()-ed because of the ethernet_any logic.

BUG=b:161454972
TEST=shill unit tests, jetstream cq tests no longer fail

Change-Id: I437328b72eb982f5246df6ee1b9b491aba14c2b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2309629
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>