crash: Add --delete_crashes flag to crash_sender.
When we upload crashes from tast tests, we will want to be able to leave
the crash files around so that tast can still save them and upload them
to stainless.

The --delete_crashes flag defaults to true, to maintain existing
behavior.

When crash_sender runs with --delete_crashes=false, it'll create a new
file with the same basename and the `.alreadyuploaded` extension to
indicate that this particular crash has been uploaded and shouldn't be
uploaded again.

An alternative to this flag would be to change tast to copy crash files
(rather than moving them) and run crash_sender after the copy. However,
since crash_sender only deletes and sends crashes on official builds
(and we do not intend to change this behavior), this alternative would
require additional logic on tast's side to examine the image type and
decide whether to move the files or to copy them -- effectively, tast
would need to duplicate crash_sender's logic to determine whether to
send crashes.

This change also attempts to create the `alreadyuploaded` file if we
fail to delete a .meta file, so that crash_sender doesn't uploading the
same crash repeatedly if a deletion fails (e.g. if selinux denies it the
ability to remove the files. See https://crbug.com/1060019).

BUG=chromium:1107005,chromium:1060019
TEST=tast run $DUT crash.Sender{.*,Delete}

Cq-Depend: chromium:2309189
Change-Id: Id8c0b1be264d3d923dbe113ed5eba2656903ff99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2305256
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>