dlcservice: Fix build rule dependency problem
There are a few problems with the current build rules:

1) libdlcservice_boot is not added to group("all") dependencies:
To fix, instead of keeping this intermediate library just moved its
source files as a part of dlcservice_util library. That is the only
place it has a dependency anyway.

2) utils.cc depends on dlc_service.h? :(
This is just pure wrong.

3) utils.h|cc depends on dlc.h:
Moved ToDlcIdList() function to dlc_manager.cc because that's the only
place it has been used. This allows us to remove the dlc.h dependency
from util. Having this dependency have caused the dlcservice_util
library be dependent on everything that dlc.h depends like ref count
protobufs, etc. We don't need this. We should try to keep the utils.cc
as independent as possible. This also causes all other libraries
depending on utils.cc to depend on dlc.h stuff too that we don't want.

BUG=chromium:1107001
TEST=FEATURES=test emerge-reef dlcservice

Change-Id: I81acc82c54977eab1d15c9cefa7e2e325a7c5720
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2305409
Tested-by: Amin Hassani <ahassani@chromium.org>
Auto-Submit: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>