system-proxy: Implement ShutDownProcess D-Bus API
This CL implements the D-Bus method ShutDownProcess which, depending on
the argument, can shut down the daemon or just one of the worker
processes.

This is necessary as the worker process which tunnels user traffic is
only active during the user session and the existing ShutDown method
doesn't accept arguments to differentiate between shutting down the
daemon or just a child process. ShutDown will be removed when the
browser code is updated to call the new method.

BUG=chromium:1042639, chromium:1109144
TEST=unit tests

Change-Id: Ide61873b06fc5fd0b030e765e68102ce4781255d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2308532
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>