login: clean up vmodule patterns
We noticed performance regressions on chrome function logging::MatchVlogPattern, one of the reasons is the increase in --vmodule patterns used for temporary debugging. The presence of any --vmodule entry causes the pattern matching function to run on every VLOG. So it would be best to remove the vmodule patterns whose issues have been closed.

BUG=b:77729987
BUG=chromium:851151,chromium:938973,chromium:942689,chromium:904600,chromium:917700
BUG=chromium:826982,chromium:952570,chromium:1015474,chromium:1011221,chromium:1046860
TEST=None

Change-Id: If03db54cfc80dcdaed45c6293f6fa9b53ad2eaba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2305853
Tested-by: Bin Du <dubin@google.com>
Commit-Queue: Xiaoyong Zhou <xzhou@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>