diagnostics: calculate prime number list for prime_search
Previously, a very large list of known prime numbers was used to verify
calculated prime numbers. This list of numbers took up a significant
amount of space on the rootfs (300kb).

Now, the known list of prime numbers is calculated using a sieve and
compared to the other method of calculating prime numbers. This
exercises the CPU in two different ways and ensures the results are the
same. The test is also modified to return on the first error, instead of
completing all prime calculations for simplicity.

BUG=chromium:1106852
TEST=FEATURES=test emerge-sarien diagnostics && cros deploy dut
diagnostics;
(dut) cros-health-tool diag --action=run_routine  --routine=prime_search

Change-Id: Ib6138b80fde720bed4c8df4531912ad2247652b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2308050
Reviewed-by: Paul Moy <pmoy@chromium.org>
Tested-by: Trent Begin <tbegin@chromium.org>
Commit-Queue: Trent Begin <tbegin@chromium.org>