dlcservice: Root mount file failure bug and logic change
When a DLC requires the root mount file and the file fails to be
created, that means it is an installation failure as the root mount file
will be missing when there is a hard dependency for it to exist.

Fix bug with printing the LOG as the "!" was missing. Verify by running
unittests.

BUG=none
TEST=FEATURES=test emerge-$B dlcservice dlcservice-client

Change-Id: I642f895c77c13e6fdecd2243630c1fa4ce7b20d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2314059
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Auto-Submit: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>