crash: Create .processing file when sending.
If a certain .meta file is malformed such that crash_sender always
crashes while processing it, it will get stuck and not be able to send
any more crashes (until the malformed file is removed ~6 months later
when the OS is too outdated).

Work around this by having crash_sender create a "foo.processing" file
while processing "foo.meta". If crash_sender later examines the crash
and sees the "foo.processing" file, it will remove the crash instead of
attempting to send it, since it likely crashed while attempting to send
it the first time.

There are some cases where a second send could be successful even if
crash_sender terminated during the first one: for instance, crash_sender
may CHECK-fail if it cannot connect to dbus for transient reasons, or if
the battery dies while crash_sender is running, it may not finish
processing the crash.

However, since crash_sender generally runs well after the relevant crash
happened, these transient failures should not be correlated with any
individual crash. (If a crash is caused by resource exhaustion, the
exhaustion may be remedied by the time crash_sender starts running and
processes it.) That means that this change should not cause crash_sender
to start under-reporting any particular crashes--except those that
always cause crash_sender to crash, which is the goal.

BUG=chromium:1108944
TEST=unit tests

Change-Id: Id6c59fe9206adae43529b69fe0c88bc8c7e65924
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2318406
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>