vm_tools: sommelier: Fix unaccelerated motion truncation
For X11 apps, unaccelerated motion is unfortunately usually interpreted
as an integer. This is typically done by casting (e.g. 0.99 -> 0) so we
lose a lot of motion events that get truncated.

In testing this only seems relevant to the unaccelerated values, which
are used to construct RawMotion events in Xi2 (by xwayland).

To counterract this sommelier will (for x11 apps only) round the motion
events to the next greater-magnitude value, so that motion events
(which by definition moved the cursor) will always register some
movement.

BUG=b:160430880
TEST=xinput --test-xi2

Change-Id: Ie64e0761f0ad90bba2dea5033d071a7c041ed31c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2315785
Tested-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>