dlcservice: Allow DlcService to finish/cancel specific installation
Allow |DlcService| to manage the DLC that is being installed by
update_engine. The DLC will only be cancelled if it fails to be
installed externally instead of the current approach of cancelling all
the DLCs that are transitioned into |DlcState::INSTALLING|. This is
required in order to keep multiple DLCs in |DlcState::INSTALLING| while
there is a specific DLC that update_engine is actually installing.

BUG=chromium:1052183
TEST=FEATURES=test emerge-$B dlcservice dlcservice-client

Change-Id: Ia0ab3569f60d1dc56a52c1db413af7a14ceecce6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2321269
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>