crash: Sanitize filename in ChromeCollector
ChromeCollector accepts any filename passed to it from Chrome in the
crash data.

ChromeCollector prepends the crash basename on to the file, so that's
not quite as abusable as it sounds, but it still seems better if we
insist on alphanumerics and _'s only.

Pass the input filenames through Sanitize to be sure.

BUG=None
TEST=tast ui.ChromeCrash*

Change-Id: Icf859eb382c5626fff8e8cc8630f0e2b574568bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2324522
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Auto-Submit: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>