cryptohome: Fix UserDataAuthTestThreaded.CheckAutoCleanupCallback race condition
It is possible for the disk cleanup triggered by InitializeUserDataAuth()
to run before the first time increment, causing the total (mocked) time
taken by the test to be lesser than kTimesFreeDiskSpaceCalled times the
period, and thus failing the test.

This change fixes that by removing that first time advancement and adjusting
the time limit.

BUG=chromium:1109598
TEST=50k times UserDataAuthTestThreaded.CheckAutoCleanupCallback

Change-Id: I6a3cf23f48dcf95e6e89dfd1a215e7764ff10098
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2324081
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Tested-by: Victor-Gabriel Savu <vsavu@google.com>
Commit-Queue: Archie Pusaka <apusaka@chromium.org>