vm_tools: Read StartVmRequest.vm when starting termina
With the changes in CL:2330111 to support DLC there are now multiple
possible install paths for termina and concierge will rely on chrome
to say where it should be loaded from. Make concierge listen to this.

This CL depends on chrome CL:2329126.

BUG=chromium:953544
TEST=Patched in CL:2329126 and started termina in component updater
and DLC configurations and checked that the right images were used.

Change-Id: I1f8b7ce6f727fa1925074d8d496bbde89373b771
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2329207
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>