installer: Speed install with big root fs and little RAM by chunking
When we run chromeos-install we install a single kernel and rootfs
from the source to two different partitions on the destination.

On systems with enough RAM, we can observe writing the second
partition finishes much faster than writing the first partition.
That's because the system buffer cache contains the contents after the
first copy and doesn't need to read the source again.

On systems where the rootfs is bigger than the available buffer cache
(RAM), the second write takes just as long as the first write (since
both have to read from the source in this case).  Specifically disk
caches typically throw out the oldest data.  In our case, that's
exactly the worst possible thing to do and we end up getting a 0%
cache hit rate for the 2nd read of the source.

We can fix this by breaking the install of the root filesystem into
chunks.  We'll install chunk 1 twice, then chunk 2 twice, etc.
Because each chunk is smaller then it will be in cache for the second
write.  For now we arbitrarily choose 4 chunks.  Why?
* Right now the root fs just barely doesn't fit, so even splitting
  into 4 chunks should give us a bit of buffer.
* Splitting into more chunks seems to give slightly worse performance
  on the handful of tests I did (maybe slightly worse readahead perf
  because we keep starting and stopping or some other per-transfer
  overhead?).  I tested 5 times at each size and took the average:
  - 1 chunk => 2 chunk: almost no difference.
  - 2 chunk => 4 chunk: took an extra 3-4 seconds compared to 1 chunk
  - 4 chunk => 8 chunk: took an extra 10 seconds compared to 1 chunk

At the moment we only support chunking when using "dd" to install.  If
someone tries to chunk to a partition that's written with UBI or NAND
we won't fail but we'll write it all during the first chunk and
writing the rest of the chunks will be no-ops.

BUG=chromium:1106818
TEST=chromeos-install

Change-Id: I9046cc562613b5928382aa0511ad27f05b591d15
Disallow-Recycled-Builds: kukui-arc-r-cq
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2330304
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Douglas Anderson <dianders@chromium.org>