init: ext-pci-drivers-allowlist: Add few basic drivers to allowlist
We are currently trying to restrict the drivers that are allowed to bind
to external PCI devices. The CL:2332890 installs a NULL allowlist, but
that essentially renders all the docks useless over PCI. The plan is to
fuzz and bless all drivers before being added to allowlist. But that is
still far away and we really do not know how it would look like. Starting
with a NULL allowlist seems very restrictive and a sudden regression for
current systems (considering currently there are no restrictions). This
patch proposes to add following entries to the whitelist temporarily
(to avoid regression / loss of functionality) until the shape of PCI
driver fuzzing becomes more clear:

1) pcieport: This really is not a driver. This is the Linux PCI
subsystem core code that deals with internal PCI bridges on every dock
and allows to enable essential services such as AER and hotplug. This is
really needed or these services won't work.

2) xhci_hcd: This is the XHCI host controller driver for controllers
present in lots of docks today, and (I think) are the primary devices on
the docks.

3) nvme: This is the PCI Express NVME host controller driver for a lot
of NVMEs available in the market.

BUG=b:153180503,b:151065827,b:140645321
TEST=Check that the 2 entries get installed in allowlist in sysfs

Change-Id: I10d3934760b7b24a815bf00f1ac1978cc30d6235
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2333357
Tested-by: Rajat Jain <rajatja@google.com>
Commit-Queue: Rajat Jain <rajatja@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>